docs(asset-tree): rewrite stories in new format (#539)
