--- src/components/AssetTree/stories/basic.md
@@ -1,20 +1,0 @@
-## Basic 
-
-<!-- STORY -->
-
-#### Usage:
-
-```typescript jsx
-import 'antd/dist/antd.css';
-
-import React from 'react';
-import { AssetTree } from '@cognite/gearbox';
-
-function ExampleComponent(props) {
-
-  return (
-    <AssetTree />
-  );
-
-}
-```
