+++ src/components/AssetTree/stories/basic.story.mdx
@@ -1,0 +1,64 @@
+import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
+import { AssetTree, AssetTreeWithoutTheme } from '../AssetTree';
+import { decorators } from './helper';
+
+<Meta title="Assets|AssetTree" decorators={decorators} />
+
+# Asset Tree
+
+Visualize parent-child relationship of assets in a tree structure. It will initially fetch all available root assets from the SDK. When a node is expanded, the children will be fetched dynamically
+
+**NOTE:** The component should have `ClientSDKProvider` as a parent component in react component tree.
+
+## Basic usage:
+
+Imports:
+
+```jsx
+import 'antd/dist/antd.css';
+
+import React from 'react';
+import { AssetTree } from '@cognite/gearbox';
+
+```
+
+<Preview>
+  <Story name="Basic Usage">
+    <AssetTree/>
+  </Story>
+</Preview>
+
+## Available props
+
+<!-- TODO -> props auto showing not working for class components -->
+<Props of={AssetTreeWithoutTheme} />
+
+### Types
+
+#### OnSelectAssetTreeParams
+
+This type describes the parameters the `onSelect` function is called with.
+The type can be imported from @cognite/gearbox:
+
+```jsx
+import { OnSelectAssetTreeParams } from '@cognite/gearbox';
+```
+
+#### Asset
+
+The Asset type documentation can be found at https://cognitedata.github.io/cognite-sdk-js/classes/asset.html
+
+#### AssetTreeStyles
+This interface defines inline CSS styles for inner elements of `AssetTree` component.
+You can override styles of following blocks:
+
+<img src="asset_tree/styling_schema.jpg" alt="List Styling" width="350px"/>
+
+The type can be imported from `@cognite/gearbox`:
+
+```jsx
+import { AssetTreeStyles } from '@cognite/gearbox';
+```
+
+
+See more details in `Custom Styles` example.
