--- src/components/AssetTree/stories/defaultExpanded.md
@@ -1,22 +1,0 @@
-## Default expanded node 
-
-<!-- STORY -->
-
-#### Usage:
-
-```typescript jsx
-import 'antd/dist/antd.css';
-
-import React from 'react';
-import { AssetTree } from '@cognite/gearbox';
-
-function ExampleComponent(props) {
-
-  return (
-    <AssetTree
-      defaultExpandedKeys={[1234]}
-    />
-  );
-
-}
-```
