+++ src/interfaces/AssetTypes.ts
@@ -64,11 +64,26 @@ export interface AssetTreeStyles {
 }
 
 export interface AssetTreeProps {
+  /**
+   * Modify the display name of an asset
+   */
   displayName?: (asset: Asset) => string;
+  /**
+   * Triggers when a node is selected
+   */
   onSelect?: (onSelect: OnSelectAssetTreeParams) => void;
+  /**
+   * List of node ids to be expanded by default
+   */
   defaultExpandedKeys?: number[];
+  /**
+   * Object that defines inline CSS styles for
+   * inner elements of the component.
+   */
   styles?: AssetTreeStyles;
+
   theme?: AnyIfEmpty<{}>;
+
   autoPagingToArrayOptions?: AutoPagingToArrayOptions;
 }
